+++ model/RoleUtilisateur.class.php
@@ -283,7 +283,7 @@ class RoleUtilisateur extends SQL
         $result = array();
         foreach ($this->query($sql, $id_u, $droit) as $line) {
             $result[] = $line['id_e'];
-        };
+        }
         return $result;
     }
 
@@ -367,5 +367,15 @@ class RoleUtilisateur extends SQL
         $droit_list = $this->getAllDroit($id_u);
         $role_list = $this->roleSQL->getAuthorizedRoleToDelegate($droit_list);
         return $this->roleSQL->getRoleLibelle($role_list);
+    }
+
+    public function getChildrenWithPermission(int $id_e_parent, int $id_u): array
+    {
+        $sql = "SELECT entite.* FROM entite " .
+            " JOIN entite_ancetre ON entite.id_e=entite_ancetre.id_e " .
+            " JOIN utilisateur_role ON entite_ancetre.id_e_ancetre=utilisateur_role.id_e " .
+            " JOIN utilisateur ON utilisateur_role.id_u = utilisateur.id_u " .
+            " WHERE entite.entite_mere=? AND utilisateur.id_u=?";
+        return $this->query($sql, $id_e_parent, $id_u);
     }
 }