+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - les annexes des actes n'étaient pas envoyées dans le bon ordre à s2low lorsqu'il y avait plusieurs fichiers avec le même nom #1271
 - Ajout d'un test lors du retour des annexes tamponnées sur s2low pour vérifier qu'elles sont dans le bon ordre #1242
+- Lorsqu'un utilisateur n'avait pas de droit entite:lecture, il pouvait récupérer la liste des entités de premier niveau #1272
 - La récupération du multi-document i-Parapheur était incomplète s'il n'y avait qu'un document supplémentaire #1273
 - Ajout d'un commentaire sur les champs to, cc et bcc des types de dossier contenant du mail sécurisé (mailsec, mailsec-bidir et studio) #1219
 - Les mails n'étaient pas toujours reçus si le nom du destinataire était accentué #1274